Move CrossSellingResult.kt to its own file
